Augmentation Base Refactor (#2117)
* refactor

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Updated augmentation base.

* Removed label & start apply/inverse

* Finished augmentaion base refactor

* container refactoring

* Added missing files

* Added ops

* Update sequential ops

* Almost there

* Fixed computation matrix computation

* Fixed randomcrop

* Fixed erasing

* almost almost

* finished

* Added missing file

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Bug and typing fix

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Added list typing

* fixed test base

* Fixed typing

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* bug fix

* Update kornia/augmentation/_2d/geometric/crop.py


* Fixed build-docs

* Fixed bfloat16 issue on torch1.13.1

* Revert the last commit

* Fixed typing

* Fixed typos

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Typo fix

+++ kornia/geometry/bbox.py
@@ -197,7 +197,7 @@ def bbox_to_mask(boxes: torch.Tensor, width: int, height: int) -> torch.Tensor:
     """
     validate_bbox(boxes)
     # zero padding the surroudings
-    mask = torch.zeros((len(boxes), height + 2, width + 2), dtype=torch.float, device=boxes.device)
+    mask = torch.zeros((len(boxes), height + 2, width + 2), dtype=boxes.dtype, device=boxes.device)
     # push all points one pixel off
     # in order to zero-out the fully filled rows or columns
     box_i = (boxes + 1).long()
